Combining Multi-Configuration project tests to one class
