--- src/test/java/school/redrover/MultiConfigurationProject4Test.java
@@ -1,55 +1,0 @@
-package school.redrover;
-
-import org.openqa.selenium.WebElement;
-import org.testng.Assert;
-import org.testng.annotations.DataProvider;
-import org.testng.annotations.Test;
-import school.redrover.model.MainPage;
-import school.redrover.model.MultiConfigurationProjectPage;
-import school.redrover.model.NewJobPage;
-import school.redrover.runner.BaseTest;
-import school.redrover.runner.TestUtils;
-
-public class MultiConfigurationProject4Test extends BaseTest {
-    @Test
-    public void testDisableMultiConfigurationProject() {
-        TestUtils.createMultiConfigurationProject(this, "MyProject", false);
-
-        WebElement enable = new MultiConfigurationProjectPage(getDriver())
-                .getDisableClick()
-                .getEnableSwitch();
-
-        Assert.assertEquals(enable.getText(),"Enable");
-    }
-
-    @Test(dependsOnMethods = {"testDisableMultiConfigurationProject"})
-    public void testEnableMultiConfigurationProject() {
-        MultiConfigurationProjectPage multiConfigurationProjectPage = new MainPage(getDriver())
-                .clickJobWebElement("MyProject");
-
-        WebElement disableProject = new MultiConfigurationProjectPage(getDriver())
-                .getEnableClick()
-                .getDisableSwitch();
-
-        Assert.assertTrue(disableProject.isDisplayed());
-    }
-
-    @DataProvider(name = "unsafeCharacter")
-    public static Object[][] provideUnsafeCharacters() {
-        return new Object[][]{{'!'}, {'@'}, {'#'}, {'$'}, {'%'}, {'^'}, {'&'},
-                {'*'}, {'['}, {']'}, {'\\'}, {'|'}, {';'}, {':'},
-                {'<'}, {'>'}, {'/'}, {'?'}};
-    }
-
-    @Test(dataProvider = "unsafeCharacter")
-    public void testVerifyAnErrorIfCreatingMultiConfigurationProjectWithUnsafeCharacterInName(char unsafeSymbol) {
-        MainPage mainPage = new MainPage(getDriver());
-        mainPage.clickNewItem();
-
-        String invalidMessage = new NewJobPage(getDriver())
-                .enterItemName(unsafeSymbol + "MyProject")
-                .getItemInvalidMessage();
-
-        Assert.assertEquals(invalidMessage, "» ‘" + unsafeSymbol + "’" + " is an unsafe character");
-    }
-}
